refactor : 파일구조 변경
