+++ src/main/java/tave/crezipsa/crezipsa/application/chat/usecase/ChatUseCase.java
@@ -1,7 +1,7 @@
 package tave.crezipsa.crezipsa.application.chat.usecase;
 
 import tave.crezipsa.crezipsa.application.chat.dto.response.GeminiChatResponse;
-import tave.crezipsa.crezipsa.application.chat.dto.response.StoryboardStructuredResponse;
+import tave.crezipsa.crezipsa.application.storyboard.dto.response.StoryboardStructuredResponse;
 
 public interface ChatUseCase {
