+++ src/main/java/tave/crezipsa/crezipsa/application/storyboard/dto/request/SaveStoryboardRequest.java
@@ -1,0 +1,9 @@
+package tave.crezipsa.crezipsa.application.storyboard.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record SaveStoryboardRequest(
+	@NotNull Long chatMessageId,
+	String title
+) {
+}
