+++ src/main/java/tave/crezipsa/crezipsa/application/storyboard/dto/response/StoryboardStructuredResponse.java
@@ -1,0 +1,9 @@
+package tave.crezipsa.crezipsa.application.storyboard.dto.response;
+
+public record StoryboardStructuredResponse(
+	String cutSummary,
+	String script,
+	String caption,
+	String time
+) {
+}
