+++ src/main/java/tave/crezipsa/crezipsa/domain/storyboard/port/StoryboardStructurerPort.java
@@ -1,6 +1,6 @@
 package tave.crezipsa.crezipsa.domain.storyboard.port;
 
-import tave.crezipsa.crezipsa.application.chat.dto.response.StoryboardStructuredResponse;
+import tave.crezipsa.crezipsa.application.storyboard.dto.response.StoryboardStructuredResponse;
 
 public interface StoryboardStructurerPort {
 	StoryboardStructuredResponse structure(String aiMessageContent);
