+++ src/main/java/tave/crezipsa/crezipsa/presentation/storyboard/StoryboardController.java
@@ -1,0 +1,4 @@
+package tave.crezipsa.crezipsa.presentation.storyboard;
+
+public class StoryboardController {
+}
